add trading functionality
Based on [kraken-api][kraken] enable basic trading functionality
of balance check and market buys. There are outstanding issues
on which the lamassu-bitstamp library (the current example to work
from) does not give guidance and need to include later, such as:

* throwing error if user tries to trade in a currency not available
* better error handling on trade, though this is mostly kraken-api's side
* include other kind of trading, not necessarily market price but e.g. limit

This last one is based on the lamassu-bitstamp's example (see the fudge
factor there), though it might be the limitation of Bitstamp API instead
of a genuine requirement. Need to clear that up.

[kraken]: https://github.com/nothingisdead/npm-kraken-api